requirements.txt
Flask
PyMuPDF
openai
numpy
scikit-learn
python-dotenv
gunicorn
Flask-Cors 